requirements.txt
absl-py==0.9.0
altgraph==0.17
astor==0.8.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
future==0.18.2
gast==0.3.3
google-pasta==0.1.8
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.18.1
onnx==1.6.0
opencv-python==4.2.0.32
pefile==2019.4.18
Pillow==7.0.0
protobuf==3.11.3
PyInstaller==3.6
pylab==0.0.2
pyusb==1.0.2
PyVISA==1.10.1
PyVISA-py==0.3.1
pywin32-ctypes==0.2.0
PyYAML==5.3
requests==2.22.0
six==1.14.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow==1.14.0
termcolor==1.1.0
tf2onnx==1.5.4
tqdm==4.42.1
typing-extensions==3.7.4.1
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.11.2
PyExecJS==1.5.1